requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
xgboost>=2.0
scipy>=1.11

requests>=2.32
beautifulsoup4>=4.12
lxml>=5.3
# selenium>=4.25  # habilite quando precisar

SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0

APScheduler>=3.10

streamlit>=1.39
plotly>=5.24
openpyxl>=3.1
reportlab>=4.2

loguru>=0.7
pytest>=8.3
black>=24.8
isort>=5.13
flake8>=7.1
python-dotenv>=1.0
scikit-learn>=1.4
joblib>=1.3
numpy>=1.26